Done with the structural change for handling multiple user types

+++ src/main/java/com/dasgupta/careercompass/bookmark/BookmarkController.java
@@ -1,7 +1,10 @@
 package com.dasgupta.careercompass.bookmark;
 
 import com.dasgupta.careercompass.job.JobDto;
+import com.dasgupta.careercompass.user.Role;
 import com.dasgupta.careercompass.user.User;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
@@ -13,6 +16,7 @@ import java.util.List;
 @RestController
 @RequestMapping("/bookmarks")
 public class BookmarkController {
+    private static final Logger log = LoggerFactory.getLogger(BookmarkController.class);
     private final BookmarkService bookmarkService;
 
     @Autowired
@@ -25,6 +29,11 @@ public class BookmarkController {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         User user = (User) authentication.getPrincipal();
 
+        log.info("The role of the requesting user is: {}", user.getRole());
+        if (user.getRole() != Role.ROLE_CANDIDATE) {
+            return ResponseEntity.badRequest().build();
+        }
+
         bookmarkService.addBookmark(user.getId(), jobId);
 
         return ResponseEntity.ok().build();
@@ -35,6 +44,11 @@ public class BookmarkController {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         User user = (User) authentication.getPrincipal();
 
+        log.info("The role of the requesting user is: {}", user.getRole());
+        if (user.getRole() != Role.ROLE_CANDIDATE) {
+            return ResponseEntity.badRequest().build();
+        }
+
         bookmarkService.removeBookmark(user.getId(), jobId);
 
         return ResponseEntity.ok().build();
@@ -44,6 +58,11 @@ public class BookmarkController {
     public ResponseEntity<List<JobDto>> getBookmarkedJobs() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         User user = (User) authentication.getPrincipal();
+
+        log.info("The role of the requesting user is: {}", user.getRole());
+        if (user.getRole() != Role.ROLE_CANDIDATE) {
+            return ResponseEntity.badRequest().build();
+        }
 
         List<JobDto> bookmarkedJobs = bookmarkService.getBookmarkedJobs(user.getId());
